refactor: extract xccy instruments and rename market to calibration
Split the cross-currency module into separate concerns:
- NEW: dal-cpp/dal/curve/xccyinstrument.hpp/cpp — CurrencyPair_,
  CrossCurrencySwap_, and their pricing helpers
- RENAMED: xccymarket.hpp/cpp → xccycalibration.hpp/cpp —
  calibration logic (CrossCurrencyMarket_, calibration spec,
  solver, diagnostics, and result types)
- Updated all #include directives in tests and examples

This improves modularity: instrument definitions are now
independent of the calibration framework, making them
reusable in other contexts without pulling in solver
dependencies.
