+++ dal-cpp/dal/curve/xccycalibration.hpp
@@ -6,27 +6,14 @@
 
 #include <map>
 #include <dal/platform/platform.hpp>
-#include <dal/currency/currency.hpp>
 #include <dal/curve/calibration.hpp>
 #include <dal/curve/curveblock.hpp>
+#include <dal/curve/xccyinstrument.hpp>
 #include <dal/math/matrix/matrixs.hpp>
 #include <dal/math/vectors.hpp>
-#include <dal/protocol/rateconvention.hpp>
-#include <dal/time/date.hpp>
 
 namespace Dal {
     class DiscountCurve_;
-
-    struct CurrencyPair_ {
-        Ccy_ domestic_;
-        Ccy_ foreign_;
-
-        CurrencyPair_();
-        CurrencyPair_(const Ccy_& domestic, const Ccy_& foreign);
-        [[nodiscard]] CurrencyPair_ Reversed() const;
-        [[nodiscard]] bool operator<(const CurrencyPair_& rhs) const;
-        [[nodiscard]] bool operator==(const CurrencyPair_& rhs) const;
-    };
 
     class CrossCurrencyMarket_ {
         Ccy_ domesticCcy_;
@@ -56,38 +43,6 @@ namespace Dal {
         [[nodiscard]] double BasisDiscountFactor(const Date_& from, const Date_& to) const;
         [[nodiscard]] double FxForward(const Date_& maturity) const;
         [[nodiscard]] double FxForward(const Date_& from, const Date_& maturity, const CollateralType_& collateral) const;
-    };
-
-    class CrossCurrencySwap_ {
-    public:
-        struct Rate_ : noncopyable {
-            virtual ~Rate_() = default;
-            virtual double operator()(const CrossCurrencyMarket_& market) const = 0;
-        };
-
-    private:
-        Date_ tradeDate_;
-        Date_ start_;
-        Date_ maturity_;
-        double marketRate_;
-        CurrencyPair_ pair_;
-        double domesticNotional_;
-        double foreignNotional_;
-        CrossCurrencyConvention_ convention_;
-
-    public:
-        CrossCurrencySwap_(const Date_& tradeDate,
-                           const Date_& start,
-                           const Date_& maturity,
-                           double marketRate,
-                           const CurrencyPair_& pair,
-                           double domesticNotional,
-                           double foreignNotional,
-                           const CrossCurrencyConvention_& convention = CrossCurrencyConvention_());
-        [[nodiscard]] String_ Name() const;
-        [[nodiscard]] pair<Date_, Date_> TimeSpan() const;
-        [[nodiscard]] double MarketRate() const { return marketRate_; }
-        [[nodiscard]] Handle_<Rate_> Precompute() const;
     };
 
     struct CrossCurrencyCalibrationDiagnostics_ {
